+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 - [eb71ba6] [Feature] Claude Sonnet 4.5
 
 > Improvements:
+- [673b563] [Improvement] Change configured models from Telegram
 - [67a3d1f] [Improvement] Update models
-- [2fbe524] [Improvement] Update xAI models
 
 > Bug Fixes:
 - [28db2f1] [BugFix] Prevent lopping /shutdown from Telegram